+++ remotion/compositions/showcases/devoxx2023/Devoxx2023.tsx
@@ -10,7 +10,7 @@ import {
 import {Talk} from '../../../types/conferences.types';
 import {Logo} from './Logo';
 import {Speakers} from './Speakers';
-import {TalkTitle} from './TalkTitle';
+import {Title} from './Title';
 import React from 'react';
 import {Details} from './Details';
 import Balloons from './Balloons';
@@ -69,7 +69,19 @@ export const Devoxx2023: React.FC<Talk> = ({
 				<Speakers speakers={speakers} />
 			</Sequence>
 			<Sequence from={25}>
-				<TalkTitle title={title} delay={ANIMATION_DELAY} />
+				<Title
+					title={title}
+					delay={ANIMATION_DELAY}
+					style={{
+						zIndex: 1,
+						position: 'absolute',
+						bottom: 130,
+						maxWidth: '90%',
+						left: 0,
+						right: 0,
+						margin: '0 auto',
+					}}
+				/>
 			</Sequence>
 			<Sequence from={40}>
 				<Details date={date} time={time} location={location} />
