+++ remotion/compositions/showcases/devoxx2023/Title.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
 	Img,
 	interpolate,
@@ -8,14 +7,15 @@ import {
 	useVideoConfig,
 } from 'remotion';
 import {loadFont} from '@remotion/google-fonts/Aldrich';
-import {Title} from '../../../design/atoms/Title';
+import React from 'react';
 
 const {fontFamily} = loadFont();
 
-export const TalkTitle: React.FC<{
+export const Title: React.FC<{
 	title: string;
+	style?: React.CSSProperties;
 	delay?: number;
-}> = ({title, delay = 0}) => {
+}> = ({title, style, delay = 0}) => {
 	const frame = useCurrentFrame();
 	const {fps} = useVideoConfig();
 
@@ -32,24 +32,19 @@ export const TalkTitle: React.FC<{
 	});
 
 	return (
-		<Title
+		<span
 			style={{
 				fontFamily,
 				color: '#fff',
 				fontSize: 40,
-				padding: '0 10px',
+				fontWeight: 'bold',
+				padding: 10,
 				textAlign: 'center',
 				textShadow: `0 0 5px #000`,
 				opacity: titleOpacity,
 				filter: `blur(${titleDeblur}px)`,
 				lineHeight: 1.4,
-				zIndex: 1,
-				position: 'absolute',
-				bottom: 130,
-				maxWidth: '90%',
-				left: 0,
-				right: 0,
-				margin: '0 auto',
+				...style,
 			}}
 		>
 			{title}
@@ -65,6 +60,6 @@ export const TalkTitle: React.FC<{
 					left: '50%',
 				}}
 			/>
-		</Title>
+		</span>
 	);
 };
