+++ remotion/compositions/showcases/lyonJS/LayerTitle.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import {loadFont} from '@remotion/google-fonts/Aldrich';
-import {Title} from '../../../design/atoms/Title';
 
 const {fontFamily} = loadFont();
 
@@ -11,20 +10,25 @@ type Props = {
 
 export const LayerTitle: React.FC<Props> = ({title, style}) => {
 	return (
-		<Title
+		<div
 			style={{
 				fontFamily,
-				position: 'relative',
 				fontWeight: 900,
 				fontSize: '38px',
+				color: 'white',
+				textAlign: 'center',
 				zIndex: '2',
+				flex: '1',
 				padding: '30px',
+				display: 'flex',
+				flexDirection: 'column',
+				alignItems: 'center',
+				justifyContent: 'center',
 				lineHeight: '52px',
-				height: '100%',
 				...style,
 			}}
 		>
 			{title}
-		</Title>
+		</div>
 	);
 };
