+++ remotion/compositions/showcases/lyonJS/Replay.tsx
@@ -2,7 +2,7 @@ import {Audio, interpolate, Sequence, staticFile} from 'remotion';
 import {Logo} from './Logo';
 import {RightTriangle} from './RightTriangle';
 import {LeftTriangle} from './LeftTriangle';
-import {TalkTitle} from './TalkTitle';
+import {Title} from './Title';
 import {AbsoluteFill} from 'remotion';
 import {BigSpeakers} from './BigSpeakers';
 import React from 'react';
@@ -41,12 +41,15 @@ export const ReplayLyonJS: React.FC<LyonJSReplayType> = ({
 			<LeftTriangle />
 			<BigSpeakers speakers={speakers} dropTop={220} />
 			<Sequence from={165}>
-				<TalkTitle
+				<Title
 					title={title}
 					style={{
 						fontSize: '48px',
+						margin: '0 auto',
 						bottom: '200px',
 						padding: '0 60px',
+						left: 0,
+						right: 0,
 					}}
 				/>
 			</Sequence>
