+++ remotion/compositions/showcases/mixit2023/Speakers.tsx
@@ -5,7 +5,7 @@ import {
 	useCurrentFrame,
 	useVideoConfig,
 } from 'remotion';
-import {TalkTitle} from './TalkTitle';
+import {Title} from './Title';
 import {Speaker} from '../../../types/conferences.types';
 import {AvatarWithCaption} from '../../../design/molecules/AvatarWithCaption';
 
@@ -57,14 +57,11 @@ export const Speakers: React.FC<{speakers: Speaker[]}> = ({speakers}) => {
 									gap: 40,
 								}}
 							>
-								<TalkTitle
+								<Title
 									title={speaker.name}
 									style={{
 										position: 'relative',
-										left: 'unset',
 										bottom: '-20%',
-										transform: 'translate(0)',
-										height: 100,
 										fontSize: '30px',
 										fontWeight: 700,
 										color: shadowColor,
