+++ remotion/compositions/showcases/touraineTech2023/BigSpeakers.tsx
@@ -1,14 +1,13 @@
 import {
 	AbsoluteFill,
-	interpolate,
 	Sequence,
 	spring,
 	useCurrentFrame,
 	useVideoConfig,
 } from 'remotion';
 import {Speaker} from './TouraineTech2023';
+import {Title} from './Title';
 import {AvatarWithCaption} from '../../../design/molecules/AvatarWithCaption';
-import {Text} from '../../../design/atoms/Text';
 
 export const BigSpeakers: React.FC<{speakers: Speaker[]; dropTop: number}> = ({
 	speakers,
@@ -24,18 +23,6 @@ export const BigSpeakers: React.FC<{speakers: Speaker[]; dropTop: number}> = ({
 		from: -600,
 		to: dropTop,
 		durationInFrames: 30,
-	});
-
-	const titleOpacity = spring({
-		frame: frame - animationDelay * 2,
-		fps,
-		from: 0,
-		to: 1,
-		durationInFrames: 60,
-	});
-
-	const titleDeblur = interpolate(frame - animationDelay * 2, [0, 20], [5, 0], {
-		extrapolateRight: 'clamp',
 	});
 	return (
 		<Sequence from={animationDelay} name="Picture">
@@ -73,22 +60,21 @@ export const BigSpeakers: React.FC<{speakers: Speaker[]; dropTop: number}> = ({
 									gap: 40,
 								}}
 							>
-								<Text
+								<Title
+									title={speaker.name}
 									style={{
-										position: 'absolute',
-										bottom: dropTop,
+										bottom: 'auto',
 										fontSize: '38px',
 										width: 350,
 										height: 200,
 										fontWeight: 700,
 										color: '#222333',
 										textShadow: `1px 1px 1px white`,
-										opacity: titleOpacity,
-										filter: `blur(${titleDeblur}px)`,
+										position: 'relative',
+										transform: 'translateY(200px)',
 									}}
-								>
-									{speaker.name}
-								</Text>
+									delay={40}
+								/>
 							</AvatarWithCaption>
 						</div>
 					);
