+++ remotion/compositions/showcases/touraineTech2023/Replay.tsx
@@ -2,7 +2,7 @@ import {Sequence} from 'remotion';
 import {Logo} from './Logo';
 import {RightTriangle} from './RightTriangle';
 import {LeftTriangle} from './LeftTriangle';
-import {TalkTitle} from './TalkTitle';
+import {Title} from './Title';
 import {AbsoluteFill} from 'remotion';
 import {Speaker} from './TouraineTech2023';
 import {Type} from './Type';
@@ -22,7 +22,7 @@ export const Replay: React.FC<ReplayType> = ({title, speakers, type}) => {
 			<LeftTriangle />
 			<BigSpeakers speakers={speakers} dropTop={250} />
 			<Sequence from={60}>
-				<TalkTitle
+				<Title
 					title={title}
 					style={{
 						fontSize: '48px',
