+++ remotion/compositions/showcases/touraineTech2023/TouraineTech2023.tsx
@@ -2,7 +2,7 @@ import {AbsoluteFill} from 'remotion';
 import {Logo} from './Logo';
 import {LeftTriangle} from './LeftTriangle';
 import {RightTriangle} from './RightTriangle';
-import {TalkTitle} from './TalkTitle';
+import {Title} from './Title';
 import {Details} from './Details';
 import {Speakers} from './Speakers';
 
@@ -32,7 +32,7 @@ export const TouraineTech2023: React.FC<TouraineTechProps> = ({
 			<RightTriangle />
 			<LeftTriangle />
 			<Speakers speakers={speakers} />
-			<TalkTitle title={title} />
+			<Title title={title} />
 			<Details date={date} time={time} location={location} />
 		</AbsoluteFill>
 	);
